split buttons into multiple files
